Fix: 다른 블록에서 중복된 이미지를 사용하는 경우 이미지가 삭제되지 않도록 수정

+++ src/hooks/useControlBlocks.jsx
@@ -34,10 +34,16 @@ const useControlBlocks = () => {
 
       setBlocks(newBlocks);
 
+      const hasNoDuplicateImage = !newBlocks.some(
+        (block, index) =>
+          oldBlock.imageURL && index !== updatedIndex && block.imageURL === oldBlock.imageURL
+      );
+
       if (
         oldBlock.imageURL &&
         ((oldBlock.tag === "img" && updatedBlock.tag !== "img") ||
-          oldBlock.imageURL !== updatedBlock.imageURL)
+          oldBlock.imageURL !== updatedBlock.imageURL) &&
+        hasNoDuplicateImage
       ) {
         deleteNoteImage(oldBlock.imageURL);
       }
